Move NodeModelComponent back into article land.
