+++ src/kit/model/index.js
@@ -17,7 +17,6 @@ export { default as ModelProperty } from './ModelProperty'
 export { default as NodeComponent } from './NodeComponent'
 export { default as NodeComponentMixin } from './NodeComponentMixin'
 export { default as NodeModel } from './NodeModel'
-export { default as NodeModelComponent } from './NodeModelComponent'
 export { default as NodeModelFactory } from './NodeModelFactory'
 export { default as NodeModelProperty } from './NodeModelProperty'
 export { default as ObjectModel } from './ObjectModel'
